requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
anyio==2.2.0
argon2-cffi==20.1.0
asgiref==3.4.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bleach==3.3.0
BLEURT @ file:///C:/Users/hgandiko/Desktop/MACHINE%20LEARNING/bleurt
blis==0.7.4
boto3==1.17.40
botocore==1.20.40
cached-property==1.5.2
cachetools==5.0.0
catalogue==2.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.12
claucy==0.0.1.991
click==7.1.2
colorama==0.4.4
cryptography==36.0.2
cycler==0.10.0
cymem==2.0.5
Cython==0.29.24
datasets==1.18.4
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.4
dj-rest-auth==2.2.4
Django==3.2.5
django-allauth==0.50.0
django-cors-headers==3.11.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
en-core-web-sm==3.0.0
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.0.12
fire==0.4.0
flatbuffers==2.0
frozenlist==1.3.0
fsspec==2021.6.1
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
huggingface-hub==0.4.0
idna==2.10
importlib-metadata==4.11.2
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyter-packaging==0.7.12
jupyter-server==1.4.1
jupyterlab==3.0.9
jupyterlab-pygments==0.1.2
jupyterlab-server==2.3.0
jupyterlab-widgets==1.0.0
jury==2.1.5
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lemminflect==0.2.1
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.4.0
mistune==0.8.4
multidict==6.0.2
multiprocess==0.70.12.2
murmurhash==1.0.5
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==2.6.3
nltk==3.6.7
notebook==6.2.0
nubia-score==0.1.5
numpy==1.21.5
oauthlib==3.2.0
openpyxl==3.0.7
opt-einsum==3.3.0
packaging==20.9
pandas==1.3.5
pandocfilters==1.4.3
parso==0.8.1
pathy==0.4.0
pickleshare==0.7.5
Pillow==8.1.0
plotly==5.3.1
preshed==3.0.5
prometheus-client==0.9.0
prompt-toolkit==3.0.16
protobuf==3.19.4
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.7.3
Pygments==2.8.0
PyJWT==2.3.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python3-openid==3.2.0
pytorch-pretrained-bert==0.6.2
pytorch-transformers==1.2.0
pytz==2021.1
pywin32==300
pywinpty==0.5.7
PyYAML==6.0
pyzmq==22.0.3
rank-bm25==0.2.1
regex==2021.11.10
requests==2.25.1
requests-oauthlib==1.3.1
responses==0.18.0
rouge-score==0.0.4
rsa==4.8
s3transfer==0.3.6
sacremoses==0.0.43
scikit-learn==0.24.1
scipy==1.6.1
seaborn==0.11.1
Send2Trash==1.5.0
sentencepiece==0.1.95
six==1.15.0
sklearn==0.0
smart-open==3.0.0
sniffio==1.2.0
spacy==3.0.3
spacy-legacy==3.0.1
sqlparse==0.4.1
srsly==2.4.0
tenacity==8.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
tf-estimator-nightly==2.8.0.dev2021122109
tf-slim==1.1.0
thinc==8.0.1
threadpoolctl==2.1.0
tokenizers==0.10.1
torch==1.7.1+cu110
torchaudio==0.7.2
torchvision==0.8.2+cu110
tornado==6.1
tqdm==4.63.0
traitlets==5.0.5
transformers==4.3.3
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.3
validators==0.18.2
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
wget==3.2
whitenoise==6.0.0
widgetsnbextension==3.5.1
wrapt==1.14.0
xlrd==2.0.1
xxhash==2.0.2
yarl==1.7.2
zipp==3.4.0
